Add files via upload
INTRODUCTION: 
The UK government gathers and releases extensive information about road incidents across the country (typically once a year). This project presents 3 datasets each containing details of the UK road accidents, casualties and vehicles recorded in the year 2019. The records data covers, but is not limited to, geographic locations, light and weather conditions, vehicle types and ages, severity, age and number of casualties, and vehicle manoeuvres, however only personal injury accidents on public roads that were reported to the police and then recorded using the STATS19 accident reporting form are included in the data. 
The goal of the project is to analyse the given data, discover the main elements associated with fatal road accidents and recommend to government agencies ways to minimize the rate,  impact and severity of road traffic accidents in the UK.
 The targeted data will be prepared and visualised using various graphs to analyse the causes and predict the likelihood of accidents based on various factors such as the driver's age, speed limit, vehicle class, season, weather conditions, day of the week, time, and location of accidents, among others.